requirements.txt
requests
beautifulsoup4
openai
langchain
chromadb
trafilatura
transformers
datasets
peft
tensorboard
trl
tiktoken
InstructorEmbedding
sentence_transformers
bitsandbytes
wandb
xformers
ctransformers[gptq] # for exllama support
auto-gptq
colorama

# For scratch model
datasets
matplotlib
einops
tensorflow_datasets
tensorflow_text